Fixed quarter pie indicator. Code refactored. Updated README.

+++ library/src/main/java/com/pletikosa/indicators/consts/Defaults.java
@@ -4,12 +4,12 @@ public class Defaults {
 
     public static final int NO_VALUE = -1;
 
-    public static final float DEFAULT_MIN_VALUE = 0;
-    public static final float DEFAULT_MAX_VALUE = 100;
+    public static final int DEFAULT_MIN_VALUE = 0;
+    public static final int DEFAULT_MAX_VALUE = 100;
     public static final int DEFAULT_ANIM_DURATION = 500;
 
-    public static final float PIE_MAX_ANGLE = 360f;
-    public static final float HALF_PIE_MAX_ANGLE = 180f;
-    public static final float QUARTER_PIE_MAX_ANGLE = 90f;
+    public static final int PIE_MAX_ANGLE = 360;
+    public static final int HALF_PIE_MAX_ANGLE = 180;
+    public static final int QUARTER_PIE_MAX_ANGLE = 90;
 
 }
